--- src/components/HomePage/BiggestContest/BiggestContest.jsx
@@ -1,37 +1,0 @@
-
-
-const BiggestContest = () => {
-    return (
-        <section className="bg-[#5900f4f9]" >
-
-            <article className='flex gap-3 bg-[#00000055] py-10 px-14 w-max rounded-r-[60px]'>
-                <div className=" flex flex-col gap-3">
-                    <img className="w-[50px] bg-white rounded-lg animate-top-to-bottom" src="/src/assets/company/838020_file.svg" alt="" />
-                    <img className="w-[50px] bg-white rounded-lg animate-top-to-bottom" src="/src/assets/company/838020_file.svg" alt="" />
-                    <img className="w-[50px] bg-white rounded-lg animate-top-to-bottom" src="/src/assets/company/838020_file.svg" alt="" />
-                </div>
-
-                <div className="flex flex-col gap-3">
-                   
-                    <img className="w-[50px] bg-white rounded-lg animate-bottom-to-top" src="/src/assets/company/838020_file.svg" alt="" />
-                    <img className="w-[50px] bg-white rounded-lg animate-bottom-to-top" src="/src/assets/company/838020_file.svg" alt="" />
-                    <img className="w-[50px] bg-white rounded-lg animate-bottom-to-top" src="/src/assets/company/838020_file.svg" alt="" />
-                </div>
-
-            </article>
-            <article></article>
-            <article></article>
-
-        </section>
-    );
-};
-
-export default BiggestContest;
-
-
-
-
-
-
-
-
