+++ src/components/HomePage/HomePage.jsx
@@ -3,19 +3,19 @@ import BiggestContest from "./BiggestContest/BiggestContest";
 import CategoryButtons from "./CategoryButtons/CategoryButtons";
 import CompanyCarousel from "./CompanyCarousel/CompanyCarousel";
 import FeaturedCompanies from "./FeaturedCompanies/FeaturedCompanies";
-import Home from "./Home/Home";
+import HomeSearch from "./HomeSearch/HomeSearch";
 import PracticeCustomizedMockInterview from "./PracticeCustomizedMockInterview/PracticeCustomizedMockInterview";
 
 
 const HomePage = () => {
     return (
         <div>
-            <Home/>
-            <PracticeCustomizedMockInterview/>
-            <CategoryButtons/>
-            <CompanyCarousel/>
-            {/* <FeaturedCompanies/> */}
-            <BiggestContest/>
+            <HomeSearch />
+            <PracticeCustomizedMockInterview />
+            <CategoryButtons />
+            <CompanyCarousel />
+            <FeaturedCompanies />
+            <BiggestContest />
         </div>
     );
 };
