+++ src/pages/HomePage/CompanyCarousel/CompanyCarousel.jsx
