+++ src/pages/HomePage/Home.jsx
@@ -1,6 +1,13 @@
 // import React from "react";
 
-import HomePage from "../../components/HomePage/HomePage";
+
+import BiggestContest from "./BiggestContest/BiggestContest";
+import CategoryButtons from "./CategoryButtons/CategoryButtons";
+import CompanyCarousel from "./CompanyCarousel/CompanyCarousel";
+import FeaturedCompanies from "./FeaturedCompanies/FeaturedCompanies";
+import HomeSearch from "./HomeSearch/HomeSearch";
+import PracticeCustomizedMockInterview from "./PracticeCustomizedMockInterview/PracticeCustomizedMockInterview";
+
 
 const Home = () => {
   return (
@@ -8,9 +15,14 @@ const Home = () => {
       {/* here add all section of home page,
       every section import here. create home page section inside the home page folder
       */}
- 
- <HomePage/>
-    
+
+
+      <HomeSearch />
+      <PracticeCustomizedMockInterview />
+      <CategoryButtons />
+      <CompanyCarousel />
+      <FeaturedCompanies />
+      <BiggestContest />
     </div>
   );
 };
